chore - optimizing product component implementing lazy loading to reduce bundle size
